ATIVIDADE: 	- atualização da medalha do usuário que acessou uma aula / realizou uma atividade (DAOAula; DAOAtividade) 	- ao inserir/atualizar uma medalha, é realizado update em usuario para atualizar medalhas 	- alterado o tipo de dado (Integer para Double) do atributo pontuacaoNecessaria de Medalha

+++ src/java/negocio/entidade/Medalha.java
@@ -13,7 +13,7 @@ public class Medalha {
 
     private Integer id;
     private String nome;
-    private Integer pontuacaoNecessaria;
+    private Double pontuacaoNecessaria;
 
     public Integer getId() {
         return id;
@@ -31,11 +31,11 @@ public class Medalha {
         this.nome = nome;
     }
 
-    public Integer getPontuacaoNecessaria() {
+    public Double getPontuacaoNecessaria() {
         return pontuacaoNecessaria;
     }
 
-    public void setPontuacaoNecessaria(Integer pontuacaoNecessaria) {
+    public void setPontuacaoNecessaria(Double pontuacaoNecessaria) {
         this.pontuacaoNecessaria = pontuacaoNecessaria;
     }
 }
